% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{categ}
\alias{categ}
\title{Shows a vector with the categories of the column of a data frame}
\usage{
categ(x, y)
}
\arguments{
\item{x}{A data frame}

\item{y}{column of the data frame}
}
\value{
A character vector.
}
\description{
Shows a vector with the categories of the column of a data frame
}
\examples{
cat<-data.frame(col1=c("cat1","cat2","cat3","cat2","cat1"),col2=c(1,2,3,4,5))
categ(cat,"col1")
}
