% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapro-package.R
\docType{package}
\name{datapro-package}
\alias{datapro}
\alias{datapro-package}
\title{datapro: Package Data Pra-pemrosesan yang Efisien}
\description{
Package ini menyediakan berbagai fungsi untuk pra-pemrosesan data dengan fokus utama pada tahapan pembersihan data, transformasi data, dan reduksi dimensi. Package ini menawarkan beberapa pilihan metode yang bisa digunakan oleh pengguna dan terintegrasi dengan aplikasi web Shiny untuk kemudahan penggunaan.
}
\keyword{internal}
