% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invers_transformasi_data.R
\name{invers_transformasi_data}
\alias{invers_transformasi_data}
\title{Mengembalikan data yang telah ditransformasi ke skala aslinya}
\usage{
invers_transformasi_data(data, params, verbose = TRUE)
}
\arguments{
\item{data}{Sebuah Data frame, tibble, atau vektor numerik yang telah ditransformasi dan akan dikembalikan ke skala asli.}

\item{params}{Parameter yang dihasilkan dari fungsi transformasi_data dengan opsi invers_params = TRUE.
Parameter ini berisi informasi transformasi yang diperlukan untuk melakukan invers.}

\item{verbose}{Logis, apakah akan mencetak informasi tentang proses invers (default TRUE).}
}
\value{
Data dalam format yang sama dengan input, tetapi telah dikembalikan ke skala aslinya.
Jika input berupa vektor, mengembalikan vektor. Jika input berupa data frame, mengembalikan data frame.
}
\description{
Fungsi ini mengembalikan data yang telah ditransformasi menggunakan fungsi transformasi_data
kembali ke skala dan distribusi aslinya. Fungsi ini memerlukan parameter yang disimpan
saat melakukan transformasi dengan opsi invers_params = TRUE.
}
\examples{
# Membuat sampel data
data_asli <- c(1, 2, 3, 4, 5, 10, 15, 20)
hasil_transform <- transformasi_data(data_asli, metode = "log", invers_params = TRUE)
data_transformed <- hasil_transform$data
params <- hasil_transform$params
data_kembali <- invers_transformasi_data(data_transformed, params)

# Membuat sampel data
df <- data.frame(a = c(1, 2, 3, 4, 5), b = c(10, 20, 30, 40, 50))
df_transform <- transformasi_data(df, kolom = c("a", "b"), metode = "boxcox", invers_params = TRUE)
df_transformed <- df_transform$data
params_transform <- df_transform$params
df_kembali <- invers_transformasi_data(df_transformed, params_transform)

}
\seealso{
\code{\link{transformasi_data}} untuk fungsi transformasi.
\code{\link{plot_transformasi}} untuk fungsi visualisasi data sebelum dan sesudah transformasi
\code{\link{normalisasi_data}} untuk fungsi menerapkan normalisasi pada dataset.
}
