% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repnas.R
\name{repnas}
\alias{repnas}
\title{Replace NAs of a a data frame by a metric, such as median or mean}
\usage{
repnas(data, metric)
}
\arguments{
\item{data}{A data frame}

\item{metric}{column of the data frame}
}
\value{
A Data Frame
}
\description{
Replace NAs of a a data frame by a metric, such as median or mean
}
\examples{
dfna<-data.frame(col1 = c("cat1", "cat2", "cat3", "cat2", "cat1", NA), col2 = c(1, NA, 3, 4, NA, 6))
repnas(dfna,"median")
}
