% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryna.R
\name{summaryna}
\alias{summaryna}
\title{If any, returns the percentage of nas en each column of a data frame}
\usage{
summaryna(x, p = FALSE)
}
\arguments{
\item{x}{A data frame}

\item{p}{The function estimates the percentage of NAs in each column.
The function will show the percentage of NAs per column.
"p" is a percentage, indicating that we want only columns with NA´s above p.}
}
\value{
A Data Frame
}
\description{
If any, returns the percentage of nas en each column of a data frame
}
\examples{
dfna<-data.frame(col1 = c("cat1", "cat2", "cat3", "cat2", "cat1", NA), col2 = c(1, NA, 3, 4, NA, 6))
summaryna(dfna, 0.02)

}
