% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tonum.R
\name{tonum}
\alias{tonum}
\title{Transforming a categorical column of a data frame into numeric}
\usage{
tonum(data, col)
}
\arguments{
\item{data}{A data frame}

\item{col}{column name to be transformed}
}
\value{
A Data Frame
}
\description{
Transforming a categorical column of a data frame into numeric
}
\examples{
cat <- data.frame(col1 = c("cat1", "cat2", "cat3", "cat2", "cat1", NA), col2 = c(1, 2, 3, 4, 5, 6))
tonum(cat, "col1")

}
