% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformasi_data.R
\name{transformasi_data}
\alias{transformasi_data}
\title{Transformasi nilai numerik dalam dataset menggunakan berbagai metode}
\usage{
transformasi_data(
  data,
  kolom = NULL,
  metode = c("log", "sqrt", "boxcox", "yeojohnson", "exp", "power"),
  lambda = NULL,
  invers_params = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{Vektor atau dataframe yang berisi data yang akan ditransformasi.}

\item{kolom}{Vektor karakter dari nama kolom yang akan ditransformasi. Jika NULL (default), semua kolom numerik akan ditransformasi.}

\item{metode}{Metode transformasi yang tersedia: "log", "sqrt", "boxcox", "yeojohnson", "exp", atau "power".}

\item{lambda}{Parameter untuk transformasi Box-Cox, Yeo-Johnson atau power (default = 2 untuk power)}

\item{invers_params}{Jika TRUE, mengembalikan parameter untuk transformasi invers (default = FALSE)}

\item{verbose}{Logis, apakah akan mencetak informasi detail (default TRUE)}
}
\value{
Jika invers_params = FALSE, mengembalikan data yang telah ditransformasi secara invisible.
Jika invers_params = TRUE, mengembalikan list dengan komponen data dan params.
}
\description{
Fungsi ini menyediakan beberapa metode untuk transformasi pada data numerik, baik sebagai vektor
maupun sebagai kolom dalam data frame, yang umum digunakan dalam analisis statistik.
}
\examples{
# Membuat contoh data
x <- c(1, 2, 3, 4, 5)
x_transform <- transformasi_data(x, metode = "log")

# Membuat contoh data
df <- data.frame(a = c(1, 2, 3, 4, 5), b = c(10, 20, 30, 40, 50))
df_transform <- transformasi_data(df, kolom = c("a", "b"), metode = "log")

}
\seealso{
\code{\link{plot_transformasi}} untuk fungsi visualisasi data sebelum dan sesudah transformasi.
\code{\link{invers_transformasi_data}} untuk fungsi mengembalikan data yang ditransformasi ke skala aslinya.
\code{\link{normalisasi_data}} untuk fungsi menerapkan normalisasi pada dataset.
\code{\link{reduksi_dimensi}} untuk fungsi melakukan reduksi dimensi pada dataset.
}
